s3-bucket-size/requirements.txt
boto3
hurry.filesize
